Consolidate HTTP implementation
The interface exposed by our HTTP implementation is tiny: just
around 100 lines, so move have everything in the same file.

Join all bits of documentation to have a single markdown file.
