--- include/measurement_kit/http/headers.hpp
@@ -1,19 +1,0 @@
-// Part of measurement-kit <https://measurement-kit.github.io/>.
-// Measurement-kit is free software. See AUTHORS and LICENSE for more
-// information on the copying conditions.
-
-#ifndef MEASUREMENT_KIT_HTTP_HEADERS_HPP
-#define MEASUREMENT_KIT_HTTP_HEADERS_HPP
-
-#include <iosfwd>
-#include <map>
-#include <string>
-
-namespace mk {
-namespace http {
-
-typedef std::map<std::string, std::string> Headers;
-
-} // namespace http
-} // namespace mk
-#endif
