+++ src/http/client.cpp
@@ -2,7 +2,7 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-#include <measurement_kit/http/client.hpp>
+#include <measurement_kit/http.hpp>
 #include "src/http/request.hpp"
 
 namespace mk {
