+++ src/http/include.am
@@ -6,9 +6,3 @@ src_http_libhttp_la_SOURCES = \
 	src/http/response_parser.cpp
 
 libmeasurement_kit_la_LIBADD += src/http/libhttp.la
-
-httpincludedir = $(includedir)/measurement_kit/http
-httpinclude_HEADERS = \
-    include/measurement_kit/http/client.hpp \
-    include/measurement_kit/http/headers.hpp \
-    include/measurement_kit/http/response.hpp
